Move message.py to __init__.py
