+++ plugwise/messages/responses.py
@@ -8,7 +8,7 @@ from plugwise.constants import (
     MESSAGE_SMALL,
 )
 from plugwise.exceptions import ProtocolError
-from plugwise.message import PlugwiseMessage
+from plugwise.messages import PlugwiseMessage
 from plugwise.util import (
     DateTime,
     Float,
